FIX: Makefile
Normal expectaction of the user is that when a project contains a Makefile,
then that makefile is the one that should work.

This is not the case of coq-contribs.

None of the Makefiles are usable.
One must first perform a strange step such as:

    $COQBIN/coq_makefile -f Make -o Makefile

Only then they can be used:

    make

This commit fixes the problem.